+++ generator/dictionaries/finder.py
