+++ generator/helpers/terminal.py
@@ -1,4 +1,4 @@
-from dictionaries import finder
+from generator.dictionaries import finder
 
 
 class Results:
@@ -13,10 +13,12 @@ class Results:
         (bin_first_place, bin_first_count), (bin_second_place, bin_second_count) = counted_bin.most_common(2)
 
         print('-'*30)
-        print(f'\nHexagram {finder.transform_king_wen(hex_first_place)} was found {hex_first_count} times')
-        print(finder.find_king_wen_meaning(finder.transform_king_wen(hex_first_place)))
-        print(f'\nHexagram {finder.transform_king_wen(hex_second_place)} was found {hex_second_count} times')
-        print(finder.find_king_wen_meaning(finder.transform_king_wen(hex_second_place)))
+        king_wen_first_place = finder.transform_king_wen(hex_first_place)
+        king_wen_second_place = finder.transform_king_wen(hex_second_place)
+        print(f'\nHexagram {king_wen_first_place} was found {hex_first_count} times')
+        print(finder.find_king_wen_meaning(king_wen_first_place))
+        print(f'\nHexagram {king_wen_second_place} was found {hex_second_count} times')
+        print(finder.find_king_wen_meaning(king_wen_second_place))
 
         print(f'\nTrigram {tri_first_place} was found {tri_first_count} times')
         print(finder.find_trigram_meaning(tri_first_place))
